✨ Feat: 모바일 홈화면 섹션 설정
